refactor: split HA methods from afxdp.rs into afxdp/ha.rs
Move 7 Coordinator HA methods + 4 tests into a dedicated ha.rs
submodule (475 lines). afxdp.rs goes from 8683 to 8219 lines.

Moved: update_ha_state, enqueue_apply_ha_state, wait_for_ha_state_apply,
export_owner_rg_sessions, ha_groups, last_cache_flush_at,
upsert_synced_session, delete_synced_session.

Pure refactor — no behavior changes.

Closes #363
